Testing PID turn

+++ src/shared/PID.cpp
@@ -5,12 +5,12 @@
 PID::PID(float kP, float kI, float kD, bool feedForward) :
   kP(kP), kI(kI), kD(kD), feedForward(feedForward), output(1) {}
 
-void PID::set(int setPoint){
+void PID::set(float setPoint){
   this->setPoint = setPoint;
 }
 
-int PID::run(int value){
-  int error = setPoint - value;
+float PID::run(float value){
+  float error = setPoint - value;
   iTerm += error;
   //the plus equals is because this is a velocity PID
   if (feedForward){
